coinrpc
local do arquivo coinrpc corrigido
